requirements.txt
aniso8601==8.1.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-decouple==3.4
python-dotenv==0.15.0
pytz==2020.5
six==1.15.0
Werkzeug==1.0.1